Rename dev component, separate rooflineHeatmap
- moved roofline helper functions to utils
